chore: add selectors for main and time dimensions
